requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0
pygame>=2.5.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
